+++ src/auth/interfaces/auth-module-options.interface.ts
@@ -22,6 +22,14 @@ export interface AuthContext {
   refreshTokenName: string;
 }
 
+/**
+ * How session creation is done. Default is create session for every login without reusing old/existing session.
+ */
+export enum SessionCreationPolicy {
+  ALWAYS_NEW,
+  REUSE_DEVICE,
+}
+
 export interface AuthModuleOptions {
   /**
    * Secret key for signing Access Tokens
@@ -227,4 +235,10 @@ export interface AuthModuleOptions {
    * forcing the user to undergo the standard OTP verification flow if they hadn't verified previously via OTP.
    */
   forceVerificationOnGoogleLogin?: boolean;
+
+  /**
+   * Optional: Session Creation on login or Reuse if existing
+   * Default: `SessionCreationPolicy.ALWAYS_NEW` Always create on login
+   */
+  sessionCreationPolicy?: SessionCreationPolicy
 }